backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
asyncpg
psycopg2-binary
geoalchemy2
pydantic
earthengine-api
python-dotenv
apscheduler
